smart-invoice-auditor/src/ocr/processor.py
"""
OCR Processing Module

This module handles the extraction of text and structured data from invoice PDFs
using OCR technology (Tesseract or AWS Textract).
"""

import os
from typing import Dict, Any, List, Optional
import pytesseract
from PIL import Image
import pdf2image
import boto3


class OCRProcessor:
    """Base class for OCR processing"""
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """Initialize the OCR processor with configuration"""
        self.config = config or {}
    
    def process_pdf(self, pdf_path: str) -> Dict[str, Any]:
        """Process a PDF file and extract text and structured data"""
        raise NotImplementedError("Subclasses must implement this method")


class TesseractProcessor(OCRProcessor):
    """OCR processor using Tesseract"""
    
    def process_pdf(self, pdf_path: str) -> Dict[str, Any]:
        """Process a PDF file using Tesseract OCR"""
        # Convert PDF to images
        images = pdf2image.convert_from_path(pdf_path)
        
        # Extract text from each page
        results = []
        for i, img in enumerate(images):
            text = pytesseract.image_to_string(img)
            results.append({"page": i+1, "text": text})
        
        # TODO: Add structured data extraction logic
        
        return {
            "invoice_id": self._extract_invoice_id(results),
            "date": self._extract_date(results),
            "total": self._extract_total(results),
            "vendor": self._extract_vendor(results),
            "line_items": self._extract_line_items(results),
            "raw_text": "\n".join([r["text"] for r in results]),
            "pages": len(results)
        }
    
    def _extract_invoice_id(self, results: List[Dict[str, Any]]) -> str:
        """Extract invoice ID from OCR results"""
        # TODO: Implement invoice ID extraction logic
        return "INVOICE-ID-PLACEHOLDER"
    
    def _extract_date(self, results: List[Dict[str, Any]]) -> str:
        """Extract date from OCR results"""
        # TODO: Implement date extraction logic
        return "DATE-PLACEHOLDER"
    
    def _extract_total(self, results: List[Dict[str, Any]]) -> float:
        """Extract total amount from OCR results"""
        # TODO: Implement total extraction logic
        return 0.0
    
    def _extract_vendor(self, results: List[Dict[str, Any]]) -> str:
        """Extract vendor information from OCR results"""
        # TODO: Implement vendor extraction logic
        return "VENDOR-PLACEHOLDER"
    
    def _extract_line_items(self, results: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Extract line items from OCR results"""
        # TODO: Implement line items extraction logic
        return []


class TextractProcessor(OCRProcessor):
    """OCR processor using AWS Textract"""
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """Initialize the Textract processor"""
        super().__init__(config)
        self.textract_client = boto3.client('textract')
    
    def process_pdf(self, pdf_path: str) -> Dict[str, Any]:
        """Process a PDF file using AWS Textract"""
        # TODO: Implement AWS Textract processing
        
        return {
            "invoice_id": "INVOICE-ID-PLACEHOLDER",
            "date": "DATE-PLACEHOLDER",
            "total": 0.0,
            "vendor": "VENDOR-PLACEHOLDER",
            "line_items": [],
            "raw_text": "RAW-TEXT-PLACEHOLDER",
            "pages": 0
        }


def create_processor(processor_type: str = "tesseract", config: Optional[Dict[str, Any]] = None) -> OCRProcessor:
    """Factory function to create an OCR processor"""
    if processor_type.lower() == "tesseract":
        return TesseractProcessor(config)
    elif processor_type.lower() == "textract":
        return TextractProcessor(config)
    else:
        raise ValueError(f"Unsupported processor type: {processor_type}")