smart-invoice-auditor/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# OCR Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
boto3>=1.28.0  # For AWS Textract integration

# Agent Framework
langchain>=0.0.267
langchain-core>=0.0.10
langchain-community>=0.0.10
langgraph>=0.0.15
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Interface
streamlit>=1.24.0

# CLI Interface
typer>=0.9.0
rich>=13.4.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0